+++ prep/src/cmd/ci.rs
@@ -1,14 +1,14 @@
 // Copyright 2026 the Prep Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use crate::cmd::{clippy, format};
+use crate::cmd::{CargoTargets, clippy, format};
 
 /// Runs CI verification.
 ///
 /// Can be ran in `extended` mode for more thorough checks.
 ///
 /// Set `fail_fast` to `false` to run the checks to the end regardless of failure.
-pub fn run(_extended: bool, fail_fast: bool) -> anyhow::Result<()> {
+pub fn run(extended: bool, fail_fast: bool) -> anyhow::Result<()> {
     let mut errs: Vec<anyhow::Error> = Vec::new();
     let mut step = |f: fn() -> anyhow::Result<()>| -> anyhow::Result<()> {
         if let Err(e) = f() {
@@ -21,7 +21,16 @@ pub fn run(_extended: bool, fail_fast: bool) -> anyhow::Result<()> {
     };
 
     step(|| format::run(true))?;
-    step(|| clippy::run(true))?;
+
+    if extended {
+        // We need to avoid --all-targets because it will unify dev and regular dep features.
+        step(|| clippy::run(CargoTargets::Main, true))?;
+        step(|| clippy::run(CargoTargets::Auxiliary, true))?;
+    } else {
+        // Slightly faster due to shared build cache,
+        // but will miss unified feature bugs.
+        step(|| clippy::run(CargoTargets::All, true))?;
+    }
 
     if errs.is_empty() {
         Ok(())